% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\docType{data}
\name{georgia_sd}
\alias{georgia_sd}
\title{Standard deviations for the Georgia data}
\format{A data frame object with 159 observations for the 7 variables
  TotPop90_sd, PctRural_sd, PctBach_sd, PctEld_sd, PctFB_sd, PctPov_sd,
  PctBlack_sd}
\source{
Created by Madetoja (2018) based on the 1990 U.S. Census of
  Population documents
}
\usage{
data(georgia_sd)
}
\description{
Standard deviations for the attributes for the Georgia census
  data set
}
\references{
Madetoja, J. (2018). Error propagation in geographically weighted
  regression. (Doctoral dissertation, Aalto University). Manuscript in
  preparation.
}
\keyword{data}
