% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plot_confidence_intervals}
\alias{plot_confidence_intervals}
\title{Plot confidence intervals}
\usage{
plot_confidence_intervals(data, SE, max, min, point_size = 0.8, title)
}
\arguments{
\item{data}{a vector of data values}

\item{SE}{a vector of standard deviations}

\item{max}{maximum value}

\item{min}{minimum value}

\item{point_size}{the size of the points}

\item{title}{the title of the plot}
}
\value{

}
\description{
This function creates a plot for values and their 95 %
  confidence intervals ordered from left to right according to the value
}
\examples{

}
\references{
Madetoja, J. (2018). Error propagation in geographically weighted
  regression. (Doctoral dissertation, Aalto University). Manuscript in
  preparation.
}
\seealso{

}
\author{
Jaakko Madetoja
}
