% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{print_maps}
\alias{print_maps}
\title{Print maps}
\usage{
print_maps(data, spatialdataframe, print_file = TRUE, plot_original = TRUE,
  point_size = 2, inner_point_size = 0.8,
  fill_colors = rev(grey(0:20/20)), plot_border = TRUE)
}
\arguments{
\item{data}{output from the function \code{\link{epgwr_mc}}}

\item{spatialdataframe}{a SpatialPointsDataFrame or SpatialPolygonsDataFrame
object as defined in package \code{\pkg{sp}} that will be used for plotting}

\item{print_file}{TRUE (default) for creating .png images and saving them to
disk, FALSE for just showing maps in R plot device one at a time}

\item{point_size}{the size of the point when plotting SpatialPointDataFrame}

\item{inner_point_size}{the size for the inner points used for original
values}

\item{fill_colors}{vector of colors used to fill the points/polygons; default
is from white ot black (rev(grey(0:20 / 20))); first and last will be used
for original points; needs at least 10 colors (otherwise they will be
recycled)}

\item{plot_border}{TRUE (default) for plotting black borders for points;
FALSE is useful when there are lots of points}
}
\value{

}
\description{
This function creates maps for point metrics resulting from
  error propagation in geographically weighted regression
}
\examples{

}
\references{
Madetoja, J. (2018). Error propagation in geographically weighted
  regression. (Doctoral dissertation, Aalto University). Manuscript in
  preparation.
}
\seealso{
\code{\link{epgwr_mc}}, \code{\link{print_maps}}
}
\author{
Jaakko Madetoja
}
